Move getDisposalTextFromResult to root of lib folder
These aren't just used by createPncDisposalsFromResult anymore and
used by an exception generator.
