+++ packages/core/phase2/lib/createPncDisposalsFromResult/createPncDisposalByFirstAndSecondDurations.test.ts
@@ -1,10 +1,10 @@
 jest.mock("./getFirstDateSpecifiedInResult")
-jest.mock("./getDisposalTextFromResult")
+jest.mock("../getDisposalTextFromResult")
 jest.mock("./isDriverDisqualificationResult")
 
 import type { Result } from "../../../types/AnnotatedHearingOutcome"
 import createPncDisposalByFirstAndSecondDurations from "./createPncDisposalByFirstAndSecondDurations"
-import { getDisposalTextFromResult } from "./getDisposalTextFromResult"
+import { getDisposalTextFromResult } from "../getDisposalTextFromResult"
 import getFirstDateSpecifiedInResult from "./getFirstDateSpecifiedInResult"
 import isDriverDisqualificationResult from "./isDriverDisqualificationResult"
