+++ packages/core/phase2/lib/createPncDisposalsFromResult/createPncDisposalByFirstAndSecondDurations.ts
@@ -1,7 +1,7 @@
 import type { Result } from "../../../types/AnnotatedHearingOutcome"
 import type { PncDisposal } from "../../../types/PncQueryResult"
 import createPncDisposal from "./createPncDisposal"
-import { getDisposalTextFromResult } from "./getDisposalTextFromResult"
+import { getDisposalTextFromResult } from "../getDisposalTextFromResult"
 import getFirstDateSpecifiedInResult from "./getFirstDateSpecifiedInResult"
 import isAmountSpecifiedInResultValid from "./isAmountSpecifiedInResultValid"
 import isDriverDisqualificationResult from "./isDriverDisqualificationResult"
