+++ packages/core/phase2/lib/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.ts
