+++ packages/core/phase2/lib/getDisposalTextFromResult/index.ts
