+++ packages/core/phase2/lib/getDisposalTextFromResult/isFailedToAppearWarrantIssued.test.ts
