+++ packages/core/phase2/lib/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
